Change semantics of FUTEX_CMPXCH2
Doing a write+wake as part of the branch where we don't sleep is
something that user-space can already do easily enough (because it
doesn't do any special operation before it would perform a wait).
So switch that around to get something that actually makes sense for an
operation that can only be performed by the kernel.