Non-canonical memory access handling
Contrary to what I though at first, x86_64 causes a #GPF when code
accesses a non-canonical address. So going along with the usual
bad-memory-access-must-cause-a-segfault rule, decode the source
instruction to figure out the faulting address and throw an E_SEGFAULT
with its faulting address properly set when a non-canonical address is
accessed.