x86_64 #DF handling & various paging fixes
Also: Get rid of the idea that 0..2Gib is permanently mapped at -2Gib.
For one: it isn't that way in i386, so code would have to be re-written
to take advantage of that.
And second: Having everything mapped isn't a good idea at all, as I had
to learn the hard way when a stack overflow overwrote the entire kernel,
causing havoc.
Long mode now just follows the 32-bit example of only mapping physical
memory that actually contains kernel data, meaning that the initial
memory layout at -2Gib in 64-bit mode is identical to the layout at 3Gib
in 32-bit mode.